Refactor selenium-based tests
